move plotting.html_surface under plotting.surface
