+++ nilearn/plotting/__init__.py
@@ -14,7 +14,6 @@ from nilearn.plotting.find_cuts import (
 )
 from nilearn.plotting.html_connectome import view_connectome, view_markers
 from nilearn.plotting.html_stat_map import view_img
-from nilearn.plotting.html_surface import view_img_on_surf, view_surf
 from nilearn.plotting.img_comparison import (
     plot_bland_altman,
     plot_img_comparison,
@@ -45,6 +44,8 @@ from nilearn.plotting.surface import (
     plot_surf_contours,
     plot_surf_roi,
     plot_surf_stat_map,
+    view_img_on_surf,
+    view_surf,
 )
 
 __all__ = [
