+++ nilearn/plotting/surface/__init__.py
@@ -1,3 +1,4 @@
+from nilearn.plotting.surface.html_surface import view_img_on_surf, view_surf
 from nilearn.plotting.surface.surf_plotting import (
     plot_img_on_surf,
     plot_surf,
@@ -12,4 +13,6 @@ __all__ = [
     "plot_surf_contours",
     "plot_surf_roi",
     "plot_surf_stat_map",
+    "view_img_on_surf",
+    "view_surf",
 ]
