+++ nilearn/plotting/tests/test_html_surface.py
@@ -7,13 +7,13 @@ from nilearn import datasets, image
 from nilearn._utils.exceptions import DimensionError
 from nilearn.datasets import fetch_surf_fsaverage
 from nilearn.image import get_data
-from nilearn.plotting import html_surface
-from nilearn.plotting.html_surface import (
+from nilearn.plotting.js_plotting_utils import decode
+from nilearn.plotting.surface import html_surface
+from nilearn.plotting.surface.html_surface import (
     colorscale,
     get_vertexcolor,
     view_img_on_surf,
 )
-from nilearn.plotting.js_plotting_utils import decode
 from nilearn.plotting.tests.test_js_plotting_utils import (
     check_colors,
     check_html,
